refactor(toolgenerate): extract the annotation artifact package — 13 files to 8, out of the size baseline
All twelve non-test files in toolgenerate were created by one commit
(2920afa9), so the annotations_* / artifacts_* prefixes are that commit's
opinion, not independent evidence. The reference graph is, and it agrees
sharply: the five annotations_* files form a closed cluster with exactly zero
edges to the five artifacts_* files.

  annotations.go   -> GeneratePlaywrightFromAnnotations, GenerateMarkdownReport,
                      BuildIssueList  (its only callees in the package)
  report.go, visual.go -> builder  (the only users of that type)
  visual.go        -> BuildLocatorCandidates -> ExtractSummaryText, JsEscapeSingle

Outside the package only the three Handle* entry points are referenced; every
generator is package-internal. The dependency split is just as clean: the
annotation handlers need *annotation.Store and nothing else, while the artifact
handlers need *capture.Store plus internal/export and internal/tools/generate.
So the new package declares its own one-method Deps rather than borrowing
toolgenerate's five-method one — and a toolgenerate.Deps value still satisfies
it, so no caller had to be re-plumbed.

Files keep their identity, minus the now-redundant prefix:
annotations.go -> annotations/handlers.go, annotations_visual.go ->
annotations/visual.go, and so on. Every exported symbol keeps its exact name.

Tests move with their code: the annotation half of toolgenerate_test.go and
the annotation handler block of handlers_coverage_test.go both land in
annotations/annotations_test.go with their own narrow fake. The two files left
behind now match what they actually cover — dispatch validation and artifact
handlers.

succeedText and lenientUnmarshal in helpers.go had no remaining caller once the
annotation handlers left, so they are gone; the annotation package calls
mcp.SucceedText / mcp.LenientUnmarshal directly.

Verified: a harness that renders every generate artifact (Playwright visual
test, Markdown report, issue list, HAR, CSP, SRI, SARIF, PR summary, test
script) from fixed fixtures produces byte-identical output against a pristine
origin/UNSTABLE checkout.
