+++ docs/features/feature/annotated-screenshots/index.md
@@ -12,7 +12,11 @@ code_paths:
   - cmd/browser-agent/tools_analyze_annotations_hints.go
   - cmd/browser-agent/server_routes_media_draw_mode.go
   - cmd/browser-agent/tools_generate_annotations.go
-  - cmd/browser-agent/tools_generate_annotations_visual.go
+  - cmd/browser-agent/internal/toolgenerate/annotations/handlers.go
+  - cmd/browser-agent/internal/toolgenerate/annotations/visual.go
+  - cmd/browser-agent/internal/toolgenerate/annotations/report.go
+  - cmd/browser-agent/internal/toolgenerate/annotations/issues.go
+  - cmd/browser-agent/internal/toolgenerate/annotations/builder.go
   - cmd/browser-agent/annotation_store.go
   - internal/schema/analyze.go
   - internal/tools/configure/capabilities/modespecs_analyze.go
@@ -25,6 +29,7 @@ test_paths:
   - internal/annotation/store_test.go
   - cmd/browser-agent/tools_analyze_annotations_test.go
   - cmd/browser-agent/tools_generate_annotations_test.go
+  - cmd/browser-agent/internal/toolgenerate/annotations/annotations_test.go
   - scripts/smoke-tests/31-annotation-parity.sh
   - tests/extension/entry-point-parity.test.js
   - npm run smoke:annotation-parity
@@ -65,7 +70,9 @@ last_verified_date: 2026-03-05
 - `internal/annotation/store_clear.go` — `ClearAll()` resets anonymous sessions, named sessions, details, and waiters (used by `configure(what:"clear", buffer:"all")` to prevent stale replay)
 - `cmd/browser-agent/tools_analyze_annotations_handlers.go` — detail response enrichment, error correlation, LLM hints, and cross-project scope safety metadata (`projects`, `scope_ambiguous`, `scope_warning`, `filter_applied`)
 - `cmd/browser-agent/tools_analyze_annotations_hints.go` — framework-aware detail hints (`design_system`, `runtime_framework`, `error_context`)
-- `cmd/browser-agent/tools_generate_annotations_visual.go` — resilient visual test generation via locator fallback candidates (`css`, `testid`, `role`, `label`, `placeholder`, `text`)
+- `cmd/browser-agent/internal/toolgenerate/annotations/visual.go` — resilient visual test generation via locator fallback candidates (`css`, `testid`, `role`, `label`, `placeholder`, `text`)
+- `cmd/browser-agent/internal/toolgenerate/annotations/handlers.go` — the three MCP entry points (`visual_test`, `annotation_report`, `annotation_issues`) and session resolution
+- `cmd/browser-agent/internal/toolgenerate/annotations/report.go` / `issues.go` / `builder.go` — Markdown report rendering, structured issue payloads, and the shared line builder
 - `internal/schema/analyze.go` + `internal/tools/configure/capabilities/modespecs_analyze.go` — analyze annotations schema/capability metadata for `url` / `url_pattern` filters
 
 ### Tests
@@ -73,6 +80,7 @@ last_verified_date: 2026-03-05
 - `internal/annotation/store_test.go` — `TestStore_SessionTTL_Is2Hours`
 - `cmd/browser-agent/tools_analyze_annotations_test.go` — enrichment fields (`selector_candidates`, `js_framework`, `component`), error correlation, hints tests
 - `cmd/browser-agent/tools_generate_annotations_test.go` — resilient locator fallback generation tests
+- `cmd/browser-agent/internal/toolgenerate/annotations/annotations_test.go` — generator unit tests (JS escaping, locator candidates, Markdown report, issue list, Playwright script) and handler no-data/named-session paths
 - `scripts/smoke-tests/31-annotation-parity.sh` — deterministic end-to-end ingest/retrieval/generation gate with bounded retries for transient startup/no_data windows
 - `scripts/smoke-tests/annotation-parity-benchmark.sh` — repeated pass-rate benchmark with threshold enforcement
 - `scripts/smoke-test.sh` — resume-mode daemon version parity guard prevents stale-daemon false negatives in `--only/--start-from` runs
